Fix BEE2_usr_base to work with 10.1 toolflow
This fixes the BEE2 User FPGA base package so that it will work with the 10.1
toolflow.  I think that base package has not worked with the 10.1 toolflow since
commit 8d0dc7b9b608a2386f815204e3965ffd044ad870.

Through multiple combinations of working 7.1 stuff and non-working 10.1 stuff,
I eventually figured out that the problem is due to that fact that 10.1
XPS_BEE2_usr_base/system.mhs uses dcm_module version 1.00.d.  For reasons I
don't understand, BEE2 usr fpga designs built with the 10.1 toolflow work OK if
the dcm_module version is changed back to version 1.00.a (specifically the
dcm_0 instance).

The only real change that I could discern between dcm_module version 1.00.a
(OK) and 1.00.b (not OK) was a three cycle delay line for the DCM reset signal.
It seems innocuous enough, but for some reason it breaks things.  Maybe someone
with more insights into the inner workings of EDK and BORPH can explore/explain
further.

Ideally, things would be made to work with dcm_module 1.00.d because version
1.00.a is no longer available in EDK 11.1 and later.  Even though System
Generator 10.1 is the last System Generator version to support Virtex2Pro, the
later ISE/EDK versions still support that chip and it could be nice to use a
newer IDE/EDK with a netlist created by a 10.1 System Generator.