Update fast_runtime.opt to support black-boxing
System generator black boxes can be used to modularize designs and reduce the
build time of the modify-build-test cycle.  To fully benefit from them, the
developer should make sure that the corresponding *_config.m files do NOT
reference the NGC netlist.  Instead, the user should include a "pcore" block
from the BEE2 library and configure it to refer to the NGC netlist.  This will
copy the NGC netlist for the black box(es) into the "pcores" subdirectory of
the copied base system.  More info can be found in a forthcoming update to the
CASPER memo on black boxing.

This change adds an option to each base system's fast_runtime.opt file so that
the ngdbuild step will be able to locate any such NGC netlists while it is
building the System Generator part of the design.