Change latency in fft_biplex_real_2x/bi_real_unscr_2x
Change the map latency in the reorder blocks of
fft_biplex_real_2x/bi_real_unscr_2x from 0 to 1.

A latency of 0 make it hard to meet timing in demanding applications (e.g.
large number of channels and/or fast timing requirements) and precludes the use
of BRAM in the implementation of the map ROMs.  If 1 proves too contentious, it
could be changed to "bram_latency" or even exposed as its own mask parameter.