Fix Misc/armed_trigger block
The armed_trigger block contains an edge detector block that has been unlinked
from the original source, but the mask initialization function wanted to do
something that could only be done to a linked block.  Since this edge detector
is not configurable from the containing armed_trigger, the parameters were
disabled and the call to the mask initialization function was commented out.