Redo library links in xeng block
This fixes some Simulink weirdness with the xeng block.  This weirdness seems
to have been introduced in commit fbbe3cb07041e4a692d558c1bf79abc4e2837cc6.
The weirdness is that Simulink would segfault when one tried to "Look under
mask" of the "Correlator/xeng" block in casper_library.mdl (pre-split) or the
"xeng" block in casper_libraray_correlator.mdl (post-split).  This segfault
behavior was seen using the 10.1 toolflow on Windows and the 11.1 toolflow on
Linux.

The following steps were enough to avoid the segfault while the library
remained open, but this workaround did not persist through saving, closing, and
reopening the library...

1. Unlock library
2. Change demux factor in xeng mask dialog from 1 to 8
3. "Look under mask" of xeng block (OK now)
4. Close window with the "under mask" contents
5. Change demux factor in xeng mask dialog from 8 to 1
6. "Look under mask" of xeng block (still OK now)
7. Close window with the "under mask" contents
8. Save library.

What seems to have ultimately eliminated the segfault behavior are these steps
(1 through 6 are the same as above)...

1. Unlock library
2. Change demux factor in xeng mask dialog from 1 to 8
3. "Look under mask" of xeng block (OK now)
4. Close window with the "under mask" contents
5. Change demux factor in xeng mask dialog from 8 to 1
6. "Look under mask" of xeng block (still OK now)
7. Select "Link Options -> Discard/Propagate Changes" and "Discard changes" for
   each block with a "parameterized link" (i.e. block with a red link arrow),
   making sure that the mask dialog boxes of the affected blocks remained the
   same.
8. Close window with the "under mask" contents
9. Save library