Make xeng_descramble masks quieter
The mask initialization code of the xeng_descramble and xeng_descramble_4ant
blocks contained a variable assignment (ship_el_del_bits =1) without a
terminating semicolon.  The result of this assignment was then displayed in the
matlab command window every time the mask was initialized.  This change simply
adds a semicolon after the assignment to supress this extraneous display.