Make pfb_fir_real_init smarter about scaling
This makes pfb_fir_real_init smarter about scaling and bit growth.  Previously,
the output value was computed "full precision" and then scaled by
2^(nextpow2(TotalTaps)).  This was problematic since the absolute maximum gain
through the filter is nowhere near the absolute maximum gain from adding
TotalTaps arbitrary numbers together.  It also resulted in larger adders than
necessary.

The adder tree now, as before, carries all fractional bit from the multiplier
outputs, but instead of growing one bit for each level of the adder tree, the
absolute maximum gain through the filter is used to determine how many
additional non-fractional bits are needed and, thanks to the wonders of modulo
math, only uses that many non-fractional bits throughout the adder tree (with
overflow set to "Wrap"!).

After the adder tree, the sum is scaled to have only one non-fractional bit.
If the specific output type can accommodate all of the fractional bits, the
output quantization is set to "Truncate" since it is not needed in that case.

To determine the maximum gain through the filter, the pfb_coeff_gen_calc.m
script was changed such that passing in -1 for the ROM index will return the
entire set of PFB coefficients.

Finally, the pfb_fir_real_init script treats a BitWidthOut of 0 (the new
default) as meaning "autosize the output" based on full-precision fractional
bits and bit-growth-as-needed non-fractional bits.