Workaround for a Xilinx bug in caddsub_dsp48e.
Disabled "Use adder only" option in the dsp48e in caddsub_dsp48e.
According to the manual, this option saves power by powering down
unused parts of the dsp48e.

There is currently a bug in the Xilinx blockset that requires some
GCC environment variables to be unset when this option is checked.
I'm disabling it until the bug is fixed to prevent user confusion.