Add a latency to u2n block to improve timing
The u2n block performs a Box-Muller conversion to convert from uniform to
normal distriubtions.  This involves a sqrt-log lookup table and a sin/cos
lookup table.  The output of the sqrt-log lookup table is multiplied by both
outputs of the sin/cos lookup table.

When building a largish ibob design at 256 MHz, the path between the sqrt-log
lookup table and the multipliers failed to meet timing.  To solve this, 1 cycle
of latency was added to the output of the sqrt-lof lookup table as well as both
outputs of the sin/cos lookup table.

The block will still produce the same sequence for the given inputs, but it
will have one cycle more latency.  This will be noticeable when comparing
(either correlating or beamforming) the output of the new version with that
from an older version that is synchronously seeded with the same seed.  The new
version's output will lag the old verion's output by exactly one sample.