Created delay_srl block.
delay_srl implements a delay using Structural HDL in a way that is
optimized for high speed designs. It explicitly places a flip flop
before the chain of SRL16+FFs so that any long combinational path
preceding the delay_srl is not burdened with the long setup time
of an SRL.

delay_srl does not conflict with the existing (erroneously named)
delay_slr block, which appears to just be a wrapper for a Xilinx
delay block anyway.