Optimized delays in bi_real_unscr_4x.
The sync delay was originally implemented as a sync_delay block,
which includes a counter, two comparators, a mux, and a feedback
loop (with zero latency and a logical block).  For small delays,
this seems wasteful.

So, sync_delay has been replaced by a new delay_srl block for any
delay that requires less than four slices.  This threshold can be
easily tuned and is documented in a comment in the init script.

The two delays following the hilbert0 block used to be implemented
as delay_slr blocks. To help meet timing, these have been replaced
with new delay_srl blocks.