Added negate_latency parameter to mirror_spectrum.
The Negate blocks in the conjugation operation of mirror_spectrum
were not meeting timing at high speed. So, there is now an option
for a non-zero negate_latency.

For designs without dsp48s, bi_real_unscr_4x sets negate_latency
to zero to maintain backwards compatibility. For designs that do
use them, it sets the latency to one.

This affects fft_biplex_real_4x when dsp48_adders is enabled,
so i've generated a new unit test for this case.