Increase mux_latency in butterfly_direct_init.m
The latency of the four muxes in butterfly_direct was hardcoded to 1 (when the
shifting schedule is not pre-defined).  This provides one flip flop (per bit)
that can be used for register retiming.  That one flip flop can be either
absorbed into the subsequent convert block or the mux itself, but not both.
Changing the latency to 2 allows both optimizations to be achieved at the same
time, thereby improving timing.

While this change is made in the butterfly_direct_init function, mux_latency
could be added as a butterfly_direct mask parameter, but that change would
require bubbling up to the top level FFT block masks as well.