Work around Matlab's boolean type conversions
Matlab's "if" keyword evaluates an empty vector as false, but Matlab's boolean
operators && and || do not.  On the other hand, both "if" and the boolean
operators will accept [0] as false and [<non-zero>] as true.  This means that
the return value of regexp can be used as a stand-alone predicate of "if" in
all cases, but can only be used as a boolean operand when it evaluates to true.
Since there's no point in using it as a boolean operand if one knows it will
always be true, it must be "converted" to a more boolean friendly form.  The
way used here is to wrap it in a legnth() call (i.e. "length(regexp(...))").
This wll return 0 (which can be used as a boolean false) for empty and non-zero
(boolean true) for non-empty.