Add addr_bram_vacc block
This adds the addr_bram_vacc block to casper_library_accumulators.mdl.  The
addr_bram_vacc block is very similar to the simple_bram_vacc block except that
it has more internal latency to improve timing and provides "addr", "dout", and
"we" ports (instead of just dout and valid) suitable for direct interfacing
with a Xilinx bram block or a "shared bram" block from the bee2_xps library.
The overall latency from "new_acc" rising edge to "we" rising edge is two
cycles.