bus_create and bus_expand
Allows the aggregation of multiple signals into a
bus and the expansion of a bus into multiple identical
streams.

Re-implementation of the cram and uncram blocks
from the gavrt library, with additional options to
output data to the Workspace using bus_expand.
Added to the Flow Control section in the library.