Shared BRAM flexibility upgrade
The toolflow has supported a range of data port sizes for the ROACH besides
the default 32 for a while now, as well as address ranges. The mask on the
Shared BRAM has been changed to reflect these additions. The ROACH now
supports data port sizes of 8,16,32,64,128 and an address range down to
2 bits.

If the data port is smaller than 32, the word/s occupying the least significant
(of 32 from the PC) bits appear at lower addresses (in the design)  e.g for a
16 bit design  data port, bits 0-15 from the PC appear as address n in the
 design, bits 16-31 appear as address n+1 from the PC etc. Similarly, for
design data ports that are larger than 32, the least significant 32 bits in the
design correspond to the lowest address from the PC, the next 32 as the lowest
address+1 etc.