munge block
The munge block has been created to allow reordering of data
signals that have been merged into busses. The number and size
of each division is provided, along with the new order. The new
order is relative to the existing division order and is referenced
from most significant bit i.e division 0 is in the most significant
place in the existing bus, and division n-1 is in the least
significant where n is the number of divisions.