bus_expand arbitrary size/type division support
The bus_expand block allows you to split a bus
into a number of divisions. These used to have to
be of the same size. This upgrade adds the option
to split a bus into a number of divisions of arbitrary
size and type. This should help clean up many designs.

Backwards compatible