snapshot circular capture change
previously, the snapshot block might be initialised and a
stop received before a trigger. This would cause data to
be captured that did not reflect the time at which the stop
pulse was received. Fixed to only be able to stop once
capture has begun i.e ignore stop events before a trigger