snapshot performance mods
The snapshot block was not optimised for performance and caused
routing constraints such that extra latency in other modules were
required.

This commit improves performance such that an emptyish design
will compile to 380MHz+ for all parameter options with the exception
of the option allowing a delay offset in capture start. This option uses
a large (32 bit) counter followed by a comparator. This circuitry
runs at a maximum of just over 300MHz. This parameter option should
thus only be used where absolutely necessary.