Fixed support for both ZDOK inputs for the ASIAA 5GSps ADC.
Also added some preliminary ROACH2 support, basically just allows the
ROACH2 setting in the XSG core config block to be used. Has not been
tested yet!

I made a minor change to the detokenize script which makes that last
'#' match lazy. This allows conditional comments to be added to the
.mhs and .ucf files without disturbing the other matches. This was
needed since both of those files types use the '#' symbol as a comment
field.