The 5G yellow block clock rate is now the physical clock frequency applied to the board.
Note: this is the clock frequency provided over the clock in SMA connector
and is not the actual sample rate. The sample rate depends on whether you are
using the boards in one- or two- channel mode, i.e. on how many cores you are
interleaving.

At some point this will be made clear in the Wiki and the yellow block
documentation itself.