The 5GSps yellow block now passes down the ADC input frequency to the pcore.
This is passed down to the PLL/MMCM which uses it to determine whether the VCO and
PFD are within their operating ranges. An error will occur now during a build if
the specified frequency won't work with the fixed MMCM parameters. There is no
quick fix for this but at least you get an error instead of being allowed to
implement the design.

I will attempt to create a coregen script to generate appropriate PLL/MMCM
parameters given the ADC input clock frequency and requested fabric clock rate.

Note: in this commit I have also remove all DCM instantiations in favor of PLLs.