Added flip-flops and a AREA_GROUP constraint for DMUX 1:2, ROACH 2 version.
One of the ZDOK pins is traced to a input/output pad that is on the other
side of the FPGA from all the other data pins. This causes a problem occasion-
ally during place and route. To fix this I placed buffers (currently only in
the DMUX 1:2 core) before the core's output and used an AREA_GROUP constraint
for the outlying pin to place its buffer close to the pad.