ROACH2 GPIO map. Xeng redraw with user-spec'd mult
The ROACH2 GPIO mapping was incorrect. There are two 8-bit ports on later revision boards. The first is designed to connect to LEDs and is a hardware connection from the 8 green onboard LEDs. The second port is unused and available for user applications. These have now been mapped in the yellow block as "led" and "gpio", respectively, each 8 bits wide.

The X-engine core was not redrawing correctly when users specified a per-multiplier implementation. Missing slash in the init file.