Some changes to help meet timing when using the 11.x tools with ADC5G on ROACH2.
Change MAP and PAR options to try less hard (not sure why this works).

Also, added two stages of buffers on the inputs to the FIFO and manually
placed certain buffers in pcore ISD_1 (which was tightened up a bit). This
helps to position the FD's attached to the one input pad for ZDOK0 which
was accidentally placed on the other side of the chip than the others.
This will be uneccessary for the next ROACH2 rev. and this commit will
likely be reverted at that point.