OPB bus device allocation fix
Allocation of devices to OPB busses was done on imprecise adhoc basis for
ROACH and ROACH2. This led to the possibility of large designs
overflowing opb2opb bridge capacities in terms of number of devices
as well as address space.

Each block now reports how many hard-coded opb0 slots it requires when
probe_bus_usage is called. This is used in gen_xps_mod_mhs to accurately
calculate the number of opb0 devices. Also, the number of opb2opb bridges
are now pre-calculated ensuring address space and number of device
restrictions are honoured. Tested for f-engine requiring 2 opb2opb bridges
as well as QDRs and katADCs.