Added ROACH1 multicast TX support to kat tge core.
Courtesy of Ramesh Karuppusamy and Jan Wagner from MPIFR BONN.
This core is listed as "ten_Gbe_v2" in the yellowblock library. The "ten_GbE" v1 block has not been patched, nor has ROACH2's v2 (kat) core.
Transmit of multicast is supported to IP addresses that have their four most significant bits set to 0xE (that is 224.0.0.0 through 239.255.255.255), which triggers the core to modify the destination MAC to be 01:00:5e:00:00:00 + destination_IP[22:0]. That is, the lower 23 bits of the destination IP address are mapped into the lower 23 bits of MAC. From wikipedia: this means that there is ambiguity in delivering packets. If two hosts on the same subnet each subscribe to a different multicast group whose address differs only in the first 5 bits, Ethernet packets for both multicast groups will be delivered to both hosts, requiring the network software in the hosts to discard the unrequired packets.