QDR Controller for ROACH2 with hardware calibration
The software calibration uses the OPB clock to calibrate the QDR.
The required calibration must run of a 200MHz clock, which also clocks
the IDELAY_CTRL element and IODELAY elements. This QDR controller is
an update from the ROACH1 QDR controller, which uses hardware calibration.

There was also a fix in the system.ucf to specify the pin placement for ppc_irq_n,
which sometimes got placed in the QDR2 bank on the FPGA and caused VCCO conflicts.
The IO standard for the QDR's were also added.

There was also a fix to the ROACH2 infrastructure block to clock the IDELAY_CTRL
element with the SYSCLK2X clock, since dly_clk was not wired and is not
available on ROACH2.