Added PC UDP Loopback via ICMP unsearchable
The core will send a UDP packet to the PC on a unbound
UDP port. This will result in the PC echoing this UDP packet
on a ICMP unsearchable encapsulation. The udp_rx.v module will
strip the ICMP header and compare the UDP payload with the
original sent UDP payload.