twiddle_general_dsp48e 'DC spike' bug fix
Settings on c_to_ri blocks for b input and coeff_gen were switched.
Also, settings on cmult were wrong.

Included in changes to script are support for generation of a default
empty block containing nothing for storage in the library. This will
allow changes to the script to not affect the library leading to less
.mdl merge problems. Will roll this out to other blocks in future
commits.