FFT default empty block creation
All blocks draw by scripts in the FFT section of the library
have had their scripts modified so that an empty block, with no
annotation is created under certain conditions (0 FFT size,
empty coefficient string).

This allows the blocks stored in the library to be empty with
the innards being populated when sensical parameter values
are set. This has a few advantages;
1. Changes to the init script do not necessitate a redraw of the
associated library block.
2. Other library blocks containing a library block that has
changed do not need to be redrawn to ensure consistency.

This should ensure that the library .mdl files need to be changed
as little as possible, preventing merge problems and lowering
development time.

Mask changes will still require an update to the library .mdl file.
Static blocks not drawn by scripts that may contain scripted blocks
may still cause problems.