Corrected an OPB read error caused by clocking roach_infrastructure using AUX.
This was introduced in commit 82807989, which set the roach_infrastructure input clock
to be the external AUX source by default. Unfortunately that meant that unless you specifically
chose SYS (the onboard 100 MHz oscillator) the MMCM would look for a clock on AUX which may or
may not be present.

This commit uses SYS as the default setting and only switches to AUX if the user sets
the IP clock source to 'aux_clk'. All other options for clock source in the XSG block will
use SYS to clock roach_infrastructure.