replace blocks of different type
If a block has the same name, checks the type as well and replaces
the block if of a different type. Previously would just rely on the
name, leading to errors as masks may be different. If the 'ReferenceBlock'
parameter is empty (Simulink blocks) uses the previous check of just name.