Made block usable in subsystems
callback script previously used find_system(gcs, ...) to locate the
XSG block. This assumes that the ten_gbe_v2 block was instantiated
at the top level which may not always be true. Changed to
find_system(bdroot, ...) which does not care where the block is
located relative to the XSG block.