Deprecated block infrastructure
Added two scripts to allow a warning dialog to pop up warning that
a block has been deprecated. The warning for each block should only
occur once per MATLAB session and should not interrupt any compile,
simulation etc.

Deprecated blocks are to be removed when the library is updated
prior to the following CASPER workshop so collaborators have a
year to find an alternative.

The idea is to add a deprecation warning to the CopyFcn of a block
so that copying it from the library pops up the warning.
It should also be added to the mask initialization area so that
the compile or simulation of an existing design containing the
blocks to be deprecated causes a warning to be displayed.