quare_transposer input number bug fix
the init script for the square_transposer had an incomplete name
for the barrel_switcher block causing a script failure as reuse_block
could not find the specified block in the libraries.

Added default value so that script can be run from the command line.
Added logic to generate an empty block when input number = 1 for storage
in the libraries.
Replaced i and js with p and qs as Matlab treats i and js as 'special'