Fixed a bug in the epb32 to opb bridge.
A bus cycle is executed twice, because the cmnd_got signal is still
active when the slave ack. This causes slaves that traditionally uses
byte enable active [3] to control bus cycles to execute twice.

On the epb32 the byte enables are all active at the same time. This
fix will ensure that only one cycle is initiated.

Do NOT use BE[*] signals on ROACH2 for flow control.