Rename ADC16 3-wire port names and relocate
This makes the signal names in the pad report a better match to the ADC16
schematic names (to facilitate deveopment, debugging, and maintenance).  The
biggest confusion minimizer is the renumbering of chip selects from 0-3 to 1-4
so that chip select 1 in the pad report does not go to chip select 2 in the
ADC16 schematic!