Add roach2 revision info to adc16 control registers
This is done is a somewhat roundabout way.  First, the adc16_interface is given
a VHDL generic which the CASPER toolflow sets based on the ROACH2 revision
selected by the user in the adc16 yellow block.  The value of that generic is
then output from the adc16_interface component.  Wires in the system.mhs file
then connect that output to a new input of the adc16_controller.  The
adc16_controller includes the bits from that input port when the 3-wire control
register is being read.

Currently, the ROACH2 revision for which the design was compiled can be read
from the 2 least significant bits of the first location in the
adc16_controller's memory space (i.e. the 3-wire control register).