Starting to support second ADC16 card
This required additional bits in the ADC16 control registers.  These bits could
have been added in a backwards compatible way, but that would have resulted in
an awkward seeming register layout.  Since the ADC16 yellow block is still in
development, backards compatibility is not necessary so the registers were
changes in a way that is NOT backwards compatible.  Ignoring backwards
compatible also allowed additional changes to the register layout that have
been on the TODO list for some time.

Note that there is not actual support for a second ADC16 card in this commit,
but it does include much of the infrastructure required to support a second
ADC16 card.