Rename some ADC16 internal signals
Rename signals to get away from "adc0" in the names.  The only ADC16 signals
that should have "adc0" or "adc1" in their names are the 3-wire signals that
are output from the adc16_controller to the FPGA pins to the GPIO headers for
cabling to ADC in ZDOK0 or ZDOK1.  Most other signals (except adc0_clk) that
had "adc0" in their names were changed to have "adc16" in their names instead.

This also removes the requirement that the yellow block have an "adc_brd" mask
parameter (which removes another irrelevant adc0/adc1 distinction).