Include number of ADCs in an ADC16 internal register
The ADC16 controllers 3-wire control register (word 0) now includes the number
of ADC chips that the design was compiled for.  This provides an easy way for
the runtime software to determine how many ADCs can and should be calibrated.

The number of chips is currently limited to 4 or 8, but this is a software
limitation only.  In theory, one could build for any number of chips from 1 to
8.  The only complication would be for less than 3, since the line clock of the
third ADC (i.e. ADC "C" of the ADC16 board in ZDOK0) is used to provide the
fabric clock.