Fix new address decode bug in adc16_controller
The new adc16_controller that supports one or two ADC16 boards added a third
register (for the 32 delay strobe bits).  Unfortunately, this third register
was not properly decoded on writes, so writing to the second or third location
both wrote to the second location and nothing ever wrote to the third location!
This fixes that (copy/paste) problem.