Clean up ADC16's 4-way replication of delay_tap
adc_unit had a 20-bit delay tap input that was treated as four 5-bit values.
Presumably, this was to allow the delay taps for the four difference channels
of an ADC to be strobed to different values on the same clock cycle.  That
functionality is not needed and was not used (adc16_interface simply
concatentated the same 5-bit value four times).  Now adc_unit simply takes a
single 5-bit value and uses it for all four channels.  This is exactly what
happened before, but now it's expressed with less verbosity.

An ADC's four channels can be loaded with different delay tap values, just not
all on the same clock cycle (they must be loaded sequentially: set delay_tap to
value for channel 1, strobe channel 1, set delay_tap to value for channel 2,
strobe channel 2).

An ADC's four channels can be loaded with the same delay tap value on the same
clock cycle (set delay_tap to the common value and stobe all channels).  This
is a useful approach for walking through all delay tap settings during
initialization.