Make ADC16 delay strobe edge-detect double wide
The delay strobe edge detector are clocked by the fabric clock and the IODELAY
elements are loaded when the framing clock is high and the RST input is high.
The original delay strobe edge detector output a 1 fabric clock cycle wide
pulse on the rising edge of a delay strobe bit.  If this pulse, which goes to
the IODELAY RST input, ocurred during the low half of the framing clock, the
delay tap value would not be loaded as expected.

This fixes that problem by making the edge detector output pulse be 2 fabric
clock cycles wide, which is one framing clock cycle wide, which guarantees that
the delay tap value will be loaded as expected.