Remove BUFG from MMCM's feedback clock (CLKFB)
Using the BUFG in the MMCM's feedback path helps ensure phase alignment between
the MMCM's input and output clocks.  Since we have no requirement for this
phase alignment (we can remove any phase offset between clock and serial data
via the IDELAY elements), we can eliminate this BUFG.  According to the Virtex
6 clocking guide (UG362), by omiting the BUFG, the "MMCM performance should
increase since the feedback clock is not subjected to noise on the core supply
since it never passes through a block powered by this supply."

This actually undoes the change in commit 3650455.