Remove ADC16 even/odd snap ambiguity
When snaphotting via the ADC's built-in snapshot mechanism, it was not possible
to determine whether the even samples of a given channel came from the "a" or
"b" lane of that channel.  This was because the snaq_req input from the
adc16_controller was clocked using the fabric clock, so it could start snapping
on any fabric clock cycle (i.e. on either the "a" or "b" sample).  This commit
changes things so that snap req is clocked by the rising edge of the frame
clock, so that the snap will always begin with the "a" sample.

I'm still not sure whether this will end up with "a" lane data in the even
samples or the odd samples (since the snap_req edge detector will hold the
snap address counter in reset for two cycles), but at least it should be
consistent from snap to snap.