Allow ADC16 to delay a/b lanes independently
Previously, the "a" and "b" lanes of an ADC16 channel were delayed by the same
amount in their IODELAY blocks.  Using a common value for both often led to
neither lane being optimally delayed for its particular clock/data skew.  This
change allows setting the "a" and "b" delays independently.