Change master clock from ADC A to ADC D (ZDOK 0)
Based on a series of tests, it was determined that the clock from ZDOK 0 ADC D
leads to better performance in terms of a wider eye pattern as determined by
performing 10 iterations of 1K-sample snapshots for each delay tap setting.

Below is a histogram table showing number of lanes that have various widths of
the eye pattern in units of delay taps (each tap equals approximately 70
picoseconds) when using master clock from ADCs A, C, and D.  ADC B's clock
cannot be used since it does not drive a clock capable input on the FPGA.  The
number of lanes always totals 32 for a given ADC clock.

+---------+-------+-------+-------+
|         | ADC A | ADC C | ADC D |
+---------+-------+-------+-------+
| <7 taps |   11  |    2  |    0  |
|  7 raps |   14  |   13  |    9  |
|  8 taps |    6  |   13  |   14  |
|  9 taps |    1  |    4  |    8  |
+---------+-------+-------+-------+

For whatever reason using the clock from ADC A is substantially worse than
using the clock from ADC C or ADC D.  Using the clock from ADC D is
significantly better than using the clock from ADC C.