Add support for ROACH2 sync_in and sync_out GPIOs
This adds two new options to the GPIO yellow block: "ROACH2:sync_in", and
"ROACH2:sync_out".  The sync_in GPIO is an input from the vertically mounted
"J10" SMA connector on the ROACH2 board.  The sync_out GPIO is an output to the
vertically mounted "J11" SMA connector in the ROACH2 board.

The DRC support for these two GPIOs is incomplete/nonexistent.  For example,
the I/O direction *must* be set to "input" when using ROACH2:sync_in for proper
operation, but the DRC step does not enforce this.