Improve ADC16 MMCM performance
Prior to this change, the MMCM was configured with:

DIVCLK_DIVIDE = 8
CLKFBOUT_MULT_F = 16.000

For a 400 MHz input clock (i.e. line clock or a 200 MHz sample clock) this
resulted in a frequency of 50 MHz at the PFD (Phase Frequency Detector) and a
VCO frequency of 800 MHz.  Having a frequency of 50 MHz at the PFD caused the
MMCM to run in "low bandwidth mode".  Using this configuration, the reliability
of the data capture was unacceptably sub-optimal despite extensive probing
showing that the input clocks were of very high quality.

We changed the MMCM configuration to:

DIVCLK_DIVIDE = 2
CLKFBOUT_MULT_F = 5.000

and appropriately adjusted the CLKOUTn_DIVIDE values to provide the same
required output frequencies.  For a 400 MHz input clock, this led to a
frequency of 200 MHz at the PFD and a VCO frequency of 1000 MHz.  Having a
frequency of 200 MHz at the PFD caused the MMCM to run in "high bandwidth
mode".  This resulted in vastly superior performance in terms of data capture
errors.  We do not know whether this performance improvement is due to running
in high bandwidth mode or running with a higher VCO frequency (or both).

This leads to limits on the ADC line clock frequency.  Combined with the ADC
clock divider mode, this leads to limits on the ADC sample clock.

In order to be in high bandwidth mode, the frequency at the PFD must be between
135 and 450 MHz for the -1 speed grade parts on the typical ROACH2.  The VCO
frequency is limited in the -1 parts to the range 600 to 1200 MHz and the
minimum value for the CLKFBOUT_MULT_F factor is 5.  Thus, the maximum frequency
at the PFD is really limited to 240 MHz (1200/5).  To further complicate
things, the DIVCLK_DIVIDE attribute cannot be set to 3 or 4 if the input clock
frequency is greater than 315 MHz.

The combined effects of all these constraints for the -1 speed grade FPGA are
summarized here (all frequencies are in MHz):

-----------+---------------+----------+-----------------+----------+--------
Line Clock | DIVCLK_DIVIDE | PFD Freq | CLKFBOUT_MULT_F | VCO Freq | MMCM BW
-----------+---------------+----------+-----------------+----------+--------
 60-100    |       1       |   60-100 |      10.000     | 600-1000 |  LOW
100-135    |       1       |  100-135 |       8.000     | 800-1080 |  LOW
135-240    |       1       |  135-240 |       5.000     | 675-1200 |  HIGH
240-270    |       2       |  120-135 |       8.000     | 960-1080 |  LOW
270-480    |       2       |  135-240 |       5.000     | 675-1200 |  HIGH
480-500    |       5       |   96-100 |      10.000     | 960-1000 |  LOW
-----------+---------------+----------+-----------------+----------+--------

The ADC sample clock frequencies corresponding to line clock frequencies are
shown here for various numbers of inputs (all frequencies in MHz):

-----------+---------------------------------+-------------+----------
           |<---- ADC Clock Frequencies ---->|             |
Line Clock | 16 inputs | 8 inputs | 4 inputs | Recommended | Supported
-----------+-----------|----------|----------+-------------+----------
 60-100    |  30-50    |  60-100  | 120-200  |     NO      |    NO
100-135    |  50-67.5  | 100-135  | 200-270  |     NO      |    NO
135-240    | 67.5-120  | 135-240  | 270-480  |     YES     |  NOT YET
240-270    |  120-135  | 240-270  | 480-540  |     NO      |    NO
270-480    |  135-240  | 270-480  | 540-960  |     YES     |    YES
480-500    |  240-250  | 480-500  | 960-1000 |     NO      |    NO
-----------+-----------+----------+----------+-------------+----------

One further detail about the change to D=2 and M=5 is that it required setting
CLKOUT0_DIVIDE_F to 2.500.  Using a non-integer value for this meant that
CLKOUT5 was disabled.  Previously CLKOUT4 and CLKOUT5 were used to generate the
180 and 270 degree phases of the fabric clock.  These two clocks are now taken
from the inverted output of the 0 and 90 degree phases of the fabric clock.
According to Xilinx Answer AR#46398, version 13.3 (and earlier, presumably) did
not properly handle propagating the 180 degree phase to the timespec of the
CLKOUTnB pins.  The recommendation is to add an explicit constraint to
accomplish that.  Until such a constraint is implemented, users whose design
use the 180 or 270 degree clocks (e.g. designs that use QDR) *MUST* use Xilinx
tools version 14 or newer.