Fix QDR controller Verilog error
The qdrc_phy_burst_align.v file of the qdr_controller_v1_00_a pcore had an
extraneous line in it that caused a warning about using variable 'i'
uninitialized and ultimately a "multiple driver" error using EDK 13.3.

Presumably earlier tool versions were OK with this line being in there or
perhaps the model that exposed this bug was the first to use QDRs in a way that
created the conditions for this error to manifest itself.