Redesign ADC16 mux logic to avoid clock-as-logic
The ADC16 design multiplexes the "a" and "b" lanes (i.e. even and odd samples)
of each channel into one signal path.  The "a" and "b" lanes are clocked using
the "frame clock", which (for 4 inputs per ADC mode) is half the sample rate,
while the multiplexed data path is clocked using the "fabric clock", which (for
4 inputs per ADC mode) is the same as the sample clock.

Prior to this commit, the selector of the multiplexer was the frame clock
itself.  This removed any a/b ordering issues so that the multiplexed data
stream was always a0,b0,a1,b1,... and never b0,a0,b1,a1,...  Unfortunately,
the constraints needed when using a clock signal as a logic signal were not
specified so these a/b muxes contained a number of unconstrained paths.  Most
designs built to date had no problems due to these unconstrained paths, but a
rev2x8 build of adc16_test.mdl did exhibit problems (bit errors) on channels E1
(bit 5) and G3 (bits 5 and 6).  These bit errors only occurred when receiving
sample data from the analog inputs; they did not occur in any test pattern
mode.  The reason for this is that the "a" and "b" lanes are always identical
in test pattern mode, so the multiplexer timing issue never resulted in
problems when in a test pattern mode.

Generating an "unconstrained paths" report for that build showed that these
bits had the shortest OFFSET IN values of any in the whole design.  Other
designs that did not exhibit the same sort of problem had minimum OFFSET IN
values that were greater than the values for the bad bits in the problematic
build.  How the unconstrained offsets ended up was completely unpredictable;
some "lucky" builds would be OK while other "unlucky" builds would not.

Leaving these paths unconstrained would be UNacceptable as it would leave each
build of an ADC16 design susceptible to these problems, but even specifying
limits for these constraints would not truly fix the problem since that would
merely prompt the tools to report a timing error yet leave the user with no way
to fix it!  Clearly a more robust solution was required.

The more robust solution is to create a copy of the frame clock in logic rather
than using the frame clock itself.  While it is easy to create a 50% duty cycle
logic signal with half the frequency of the fabric clock, it is not quite so
trivial to create one that is "in phase" with the frame clock rather than an
inverted "copy" of it.  The "in phase" part is important to avoid the
"b0,a0,b1,a1,..." sample ordering problem.  In this context, "in phase"
actually means "with deterministic phase relationship to" (i.e. a signal that
is guaranteed to be an inverted copy of the frame clock can be just as useful
as a guaranteed non-inverted copy of the frame clock).

The method (some might say "trick") used to create the copy of the frame clock
with deterministic phase is to use the frame clock to generate a clock enable
signal for the divide-by-two counter that would otherwise create a copy of the
frame clock withOUT deterministic phase.  The clock enable signal is simply a
flip flop that initializes to "0" and is clocked to a "1" on the first rising
edge of the frame clock.  Because the frame clock and the fabric clock are
outputs from the same MMCM, the tools handle this in the desired/expected way
by propagating the PERIOD constraints to the synchronous logic.